Move y_hat estimate down a line in TRME
Want to make the redescend loops in TRME, TRME_RR the same.
In TRME_RR we need to first estimate b before updating y_hat, because
y_hat depends on b.  In TRME the order doesn't matter since both
b and y_hat depend only on QR=X and y_cleaned (which is already
updated).  Therefore, move the y_hat estimate below b in TRME and the
methods are the same.

[Issue(s): #108]