Debugging Test Failues
For not understood reasons am getting an assertion error in
test_doc_build.  there is a warning being thrown from the test_data.py
module in sandbox/io_helpers.  The test prefix here is an adjective,
not a verb, so I changed it to testing_data.py, and the related imports.

FWIW make_parkfield_mth5_from_local_data.py is not passing locally but
it is not an officially supported test anyhow, but should be looked at
when time permits, and fixed or removed.