Deprecate Process Single Decimation Level PKD Test
This test was needed early on, but it basically is a complete re-write
of the guts of process_mth5_run, which is slated to be depreacted when
we merge dev into main.

This test is accomplised already via integrated tests in synthetic,
and is also redundant given that we process all decimation levels
in test_process_parkfield_run.py.