Add handling for estimator_engine
While working on issue#80, and PR184, have noticed that processing
config defaults to estimator.engine = "RME_RR".  This is fine, but
I find I need to specify to use "RME" explicitly when there is only one
station.  So a couple fixes were added:
1. Processing class now has a validate() method.  If there is no RR
station, _and_ the estimator.engine is "RME_RR", it gets reset to
"RME".  Also added the ability to pass a kwarg to ConfigCreator instance
called estimator.  The kwarg is a dict and if "engine" is a key, it will
overwrite the estimator with the corresponding value.

The parkfield SS run test was updated to use the config_creator method.
cas04 test is usign validate()

[Issue(s): #80]