Simplify processing config initialization
While working on doc and notebooks for workshop, I realize that the
processing config always needs its stations assigned explictly by the
user before calling the pipeline,  but this information is actually
already in the kernel datset, so this explicit assignment can be
avoided, to simplify use.

To that end I sligthly modified the start of the pipeline so that
if the config doesn't have a local station, it gets the station info
from the kernel dataset.

Also updated ipynb accordingly.