Add a test to read Gary's matlab z-file
This is only a worked example, not something that is supported.
This maybe deprecated in future.
I am putting it here so that any possible dependecies on the z-file
writter used in aurora testing are archived in test folder, because
the z-file writter is going to be replaced with mt_metadata version
soon.