Add check at the run level for valid processing data
So far, each decimation level is being checked to see if there are
any valid runs, and decimation levels with no valid runs are skipped.

However, it is possible that a decimation level has some, but not all
runs valid.  In that case we need to check each run before processing it