Debugging issue #188
In this commit is a hodgepodge of experiments aimed at setting the
start/end times in mt_metadata for the synthetic data.

Many lines are from the old fix_issue_188 branch which has not been
looked at in some time, and not all are needed.

These will be cleaned up in following commits

[Issue(s): #188]