Bug Fix
There is an issue when the time intervals are incorrect,
this was handled by returning "None".
It was never expected to be encountered,
But is seems that there are mth5s with end time earlier than start time.

While this should be fixed upstream, for now in order to avoid an
exception in building the kernel_dataset, we should at least
return the correctly shaped output.

Since overlap() method is supposed to return a start_time and and
end_time, returning None is not acceptable, but returning
None, None is OK (at least structurally)

This fix is being inserted to support task: WideScale Testing

[Issue(s): #252]