Improve workaround for missing metadata
- Needed to add dummy filters for SI to MT units but,
sometimes the metadata have unknown units and sometimes they have SI units
(in the 22111 cases), so I added a check for units being unknown or not.
If units are unknown, a scaling filter with units-"unknown" is added,
otherwise SI-to-MT

This is not a general solution, as it makes assumptions about the
specific cases encountered in the IRIS dataset, BUT it does have most of
the skeleton of a general solution.

Also, the unknown units should be fixed in the metadata, not kludged in the code