Add schema to 01
- Probably would have been better to do this in two stages,
first formalize the schema, and then rename _remote_ref_type with _processing_type,
as this will break backward compatability with old 01 csvs.
- But, three reasons to press ahead: a) this is an experimental branch in the first place,
b) the change is minor and workaroundable if needed, c) the new nomenclature is formalized
in TF class in mt_metadata

In this commit I have replaced all instances of _remote_ref_type with _processing_type
that I expect to encounter moving forward.  Any leftover instances are slated for deletion.