Issue #233
This is a temporary workaround, which can also double as a long term fix.
The logic is to call the new run_summary.check_runs_are_valid(drop=True)
after the run_summary is created, but before the kernel_dataset.
In this way, runs that are all zero will not have a chance to enter process_mth5.

The only reason to not use this in general is that it requires non-lazy access to the
time series, so there is a time and processing cost.